after bully removal
