+++ src/main/java/chat/election/BullyElectionManagementService.java
@@ -1,0 +1,39 @@
+package chat.election;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+import org.quartz.*;
+import chat.common.model.ServerInfo;
+import chat.model.Constant;
+import chat.service.JSONMessageBuilder;
+import chat.service.PeerClient;
+import chat.service.Quartz;
+import chat.service.ServerState;
+import chat.election.timeout.ElectionAnswerMessageTimeoutFinalizer;
+import chat.election.timeout.ElectionCoordinatorMessageTimeoutFinalizer;
+
+import java.util.List;
+
+public class BullyElectionManagementService {
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+    private static final Logger logger = LogManager.getLogger(BullyElectionManagementService.class);
+}
