+++ src/main/java/chat/election/timeout/FastBullyCoordinatorMessageTimeoutFinalizer.java
@@ -1,4 +1,4 @@
-package chat.service.election.timeout;
+package chat.election.timeout;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -6,7 +6,7 @@ import org.quartz.DisallowConcurrentExecution;
 import org.quartz.JobExecutionContext;
 import org.quartz.JobExecutionException;
 import chat.common.model.ServerInfo;
-import chat.service.election.FastBullyElectionManagementService;
+import chat.election.FastBullyElectionManagementService;
 
 @DisallowConcurrentExecution
 public class FastBullyCoordinatorMessageTimeoutFinalizer extends MessageTimeoutFinalizer {
