+++ src/main/java/chat/handler/management/election/AnswerElectionMessageHandler.java
@@ -6,7 +6,7 @@ import org.json.simple.JSONObject;
 import chat.common.model.Protocol;
 import chat.handler.IProtocolHandler;
 import chat.handler.management.ManagementHandler;
-import chat.service.election.BullyElectionManagementService;
+import chat.election.BullyElectionManagementService;
 
 public class AnswerElectionMessageHandler extends ManagementHandler implements IProtocolHandler {
 
@@ -21,12 +21,12 @@ public class AnswerElectionMessageHandler extends ManagementHandler implements I
         System.out.println("Received answer from : " + jsonMessage.get(Protocol.serverid.toString()));
 
         // since the answer message timeout is no longer needed, stop that timeout first
-        new BullyElectionManagementService().stopWaitingForAnswerMessage(serverState.getServerInfo());
-
-        // start waiting for the coordinator message
-        new BullyElectionManagementService().startWaitingForCoordinatorMessage(
-                serverState.getServerInfo(),
-                serverState.getElectionCoordinatorTimeout());
+//        new BullyElectionManagementService().stopWaitingForAnswerMessage(serverState.getServerInfo());
+//
+//        // start waiting for the coordinator message
+//        new BullyElectionManagementService().startWaitingForCoordinatorMessage(
+//                serverState.getServerInfo(),
+//                serverState.getElectionCoordinatorTimeout());
 
     }
