+++ src/main/java/chat/heartbeat/GossipJob.java
@@ -11,8 +11,8 @@ import chat.model.Lingo;
 import chat.service.JSONMessageBuilder;
 import chat.service.PeerClient;
 import chat.service.ServerState;
-import chat.service.election.BullyElectionManagementService;
-import chat.service.election.FastBullyElectionManagementService;
+import chat.election.BullyElectionManagementService;
+import chat.election.FastBullyElectionManagementService;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -87,13 +87,13 @@ public class GossipJob implements Job {
 
                 } else {
 
-                    new BullyElectionManagementService().startElection(
-                            serverState.getServerInfo(),
-                            serverState.getCandidateServerInfoList(),
-                            serverState.getElectionAnswerTimeout());
-
-                    new BullyElectionManagementService().startWaitingForAnswerMessage(
-                            serverState.getServerInfo(), serverState.getElectionAnswerTimeout());
+//                    new BullyElectionManagementService().startElection(
+//                            serverState.getServerInfo(),
+//                            serverState.getCandidateServerInfoList(),
+//                            serverState.getElectionAnswerTimeout());
+//
+//                    new BullyElectionManagementService().startWaitingForAnswerMessage(
+//                            serverState.getServerInfo(), serverState.getElectionAnswerTimeout());
 
                 }
             }
